requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.20.0
Pillow>=8.0.0
matplotlib>=3.2.0
PyYAML>=5.3.1
torch>=1.7.0
torchvision>=0.8.0
tqdm>=4.41.0
